Color temperature slider now snaps to 100K increments
This commit updates the behavior of the temperature slider so that, as the user drags it, it snaps to the nearest 100K increment temperature value. For example, if the user drags their finger to 4570K, the slider snaps to 4600K.

To do this, we add a new function to MainViewController.m, `roundOrangeValue()`. Then, in the event handler for the slider value changed, we grab the current value of the slider, call `roundOrangeValue()` to round the value to the nearest 100K, and set the slider's position to this rounded value. The increment that we round to (100K, in this case) is controlled by the `#define ORANGE_VALUE_INCREMENT` statement at the top of `MainViewController.m`.

The benefit of this new behavior is that it makes it easier for the user to set the slider to a nice, round number. Most users probably want their temperature to be something like 3700K, rather than 3692K. With the old behavior, the slider was so sensitive that you could rarely get it exactly on a round number. Even if you could, lifting your finger off the screen would cause the slider to jump to some other ugly value.

With this new behavior, it's much easier to set the slider to a round number and keep it there. This behavior matches f.lux's behavior (at least on Mac OS X).